Rename `EntryItem` to `EntryInMatch`
